Move the clipboard handling from the Backend trait into PlatformWindow
Access to the clipboard implies access to the windowing system.
